# Proyecto/E3 (Final)/BACKEND/swagger.yaml
openapi: 3.0.0
info:
  title: Dawdle API
  description: |
    API para gestión de eventos y usuarios
    
    ## Autenticación
    La API utiliza un sistema híbrido de autenticación:
    - **JWT (JSON Web Token)**: Se devuelve en el cuerpo de la respuesta con expiración de 1 hora
    - **Cookies httpOnly**: Se establece automáticamente en las rutas `/auth/register` y `/auth/login` con las siguientes características:
      - `httpOnly: true` - No accesible desde JavaScript (protección contra XSS)
      - `secure: true` - Solo en HTTPS (producción)
      - `sameSite: 'lax'` - Protección contra CSRF
      - `maxAge: 3600000` - 1 hora (3600 segundos)
      - `path: '/'` - Disponible en toda la aplicación
    
    ## Seguridad de contraseñas
    Las contraseñas se almacenan utilizando bcrypt con **12 rounds de salt** para máxima seguridad.
    
    ## CORS
    El servidor acepta peticiones solo desde el origen configurado en `FRONT_ORIGIN` (variable de entorno).
  version: 1.0.0
  contact:
    name: Soporte Dawdle
    email: soporte@dawdle.com
servers:
  - url: http://localhost:3000
    description: Servidor local
  - url: https://dawdle-backend.onrender.com
    description: Servidor de producción
...
paths:
...
  /events:
    get:
      tags:
        - Events
      summary: Obtener eventos del usuario
      description: Retorna todos los eventos del usuario autenticado
      responses:
        '200':
          description: Lista de eventos obtenida
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventWithRelations'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Acceso denegado. Token no proporcionado."
    post:
      tags:
        - Events
      summary: Crear nuevo evento
      description: Crea un nuevo evento asociado al usuario autenticado
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventCreate'
      responses:
        '201':
          description: Evento creado exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Evento creado exitosamente"
                  event:
                    $ref: '#/components/schemas/Event'
        '400':
          description: Error de validación en los datos
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                    items:
                      type: object
                example:
                  errors:
                    - msg: "El título debe tener entre 3 y 200 caracteres"
                      param: "title"
                    - msg: "end_date no puede ser anterior a date"
                      param: "end_date"
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Acceso denegado. Token no proporcionado."
  /events/{id}:
    get:
      tags:
        - Events
      summary: Obtener evento específico
      description: Retorna un evento específico (solo dueño o admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: Evento obtenido exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventWithUser'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Acceso denegado. Token no proporcionado."
        '403':
          description: Sin permisos para ver este evento
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "No tienes permiso para ver este evento"
        '404':
          description: Evento no encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Evento no encontrado"
    put:
      tags:
        - Events
      summary: Actualizar evento
      description: Actualizar evento existente (solo dueño o admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventUpdate'
      responses:
        '200':
          description: Evento actualizado exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Evento actualizado exitosamente"
                  event:
                    $ref: '#/components/schemas/Event'
        '400':
          description: Error de validación
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: array
                example:
                  errors:
                    - msg: "end_date no puede ser anterior a date"
                      param: "end_date"
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Token expirado."
        '403':
          description: Sin permisos para editar este evento
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "No tienes permiso para editar este evento"
        '404':
          description: Evento no encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Evento no encontrado"
    delete:
      tags:
        - Events
      summary: Eliminar evento
      description: Eliminar evento (solo dueño o admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: Evento eliminado exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Evento eliminado exitosamente"
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Acceso denegado. Token no proporcionado."
        '403':
          description: Sin permisos para eliminar este evento
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "No tienes permiso para eliminar este evento"
        '404':
          description: Evento no encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Evento no encontrado"
  /events/{id}/complete:
    post:
      tags:
        - Events
      summary: Marcar evento como completado
      description: Marcar evento como completado (solo dueño)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '201':
          description: Evento marcado como completado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Evento marcado como completado"
                  completed:
                    $ref: '#/components/schemas/Completed'
        '400':
          description: El evento ya está marcado como completado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "El evento ya está marcado como completado"
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Token inválido."
        '403':
          description: Sin permisos (solo el dueño puede marcar como completado)
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "No tienes permiso para completar este evento"
        '404':
          description: Evento no encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Evento no encontrado"
    delete:
      tags:
        - Events
      summary: Desmarcar evento como completado
      description: Desmarcar evento como completado (solo dueño)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: Evento desmarcado como completado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Evento desmarcado como completado"
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Acceso denegado. Token no proporcionado."
        '403':
          description: Sin permisos (solo el dueño)
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "No tienes permiso para descompletar este evento"
        '404':
          description: El evento no está marcado como completado o no existe
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                example:
                  error: "El evento no está marcado como completado"
  /events/all:
    get:
      tags:
        - Events
      summary: Obtener todos los eventos (Admin)
      description: Retorna todos los eventos de todos los usuarios (solo admin)
      responses:
        '200':
          description: Lista de eventos obtenida
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventWithUser'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Acceso denegado. Token no proporcionado."
        '403':
          description: No tiene permisos de administrador
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Acceso denegado. Se requieren permisos de administrador."
components:
  schemas:
...
    Event:
      type: object
      properties:
        id:
          type: integer
          example: 15
        user_id:
          type: integer
          example: 2
        title:
          type: string
          example: "Estudiar Node.js"
        description:
          type: string
          example: "Repasar Express y Sequelize"
        date:
          type: string
          format: date-time
          example: "2025-11-08T14:00:00.000Z"
        end_date:
          type: string
          format: date-time
          example: "2025-11-08T16:00:00.000Z"
        all_day:
          type: boolean
          example: false
        color:
          type: string
          example: "#3498DB"
        notify:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    EventCreate:
      type: object
      required:
        - title
        - date
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 200
          example: "Estudiar Node.js"
          description: Título del evento (3-200 caracteres)
        description:
          type: string
          maxLength: 1000
          example: "Repasar Express y Sequelize"
          description: Descripción opcional del evento (máximo 1000 caracteres)
        date:
          type: string
          format: date-time
          example: "2025-11-08T14:00:00Z"
          description: Fecha y hora de inicio del evento (ISO 8601)
        end_date:
          type: string
          format: date-time
          example: "2025-11-08T16:00:00Z"
          description: "Fecha y hora de fin del evento (ISO 8601). IMPORTANTE: Debe ser mayor o igual a 'date' (validación personalizada implementada)"
        all_day:
          type: boolean
          example: false
          description: Indica si el evento dura todo el día
        color:
          type: string
          pattern: "^#[0-9A-Fa-f]{6}$"
          example: "#3498DB"
          description: Color del evento en formato hexadecimal (por ejemplo, #3498DB)
        notify:
          type: boolean
          example: true
          description: Indica si se envían notificaciones para este evento
    EventUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 200
          example: "Estudiar Node.js - Actualizado"
          description: Título del evento (3-200 caracteres)
        description:
          type: string
          maxLength: 1000
          example: "Repasar Express, Sequelize y JWT"
          description: Descripción opcional del evento (máximo 1000 caracteres)
        date:
          type: string
          format: date-time
          example: "2025-11-08T15:00:00Z"
          description: Fecha y hora de inicio del evento (ISO 8601)
        end_date:
          type: string
          format: date-time
          example: "2025-11-08T17:00:00Z"
          description: "Fecha y hora de fin del evento (ISO 8601). IMPORTANTE: Debe ser mayor o igual a 'date' (validación personalizada implementada)"
        all_day:
          type: boolean
          example: false
          description: Indica si el evento dura todo el día
        color:
          type: string
          pattern: "^#[0-9A-Fa-f]{6}$"
          example: "#E74C3C"
          description: Color del evento en formato hexadecimal (por ejemplo, #E74C3C)
        notify:
          type: boolean
          example: false
          description: Indica si se envían notificaciones para este evento
    EventWithRelations:
      allOf:
        - $ref: '#/components/schemas/Event'
        - type: object
          properties:
            Notifications:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: integer
                    example: 1
                  notif_time:
                    type: string
                    format: date-time
                    example: "2025-11-08T09:00:00.000Z"
            Completeds:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: integer
                    example: 1
                  completion_date:
                    type: string
                    format: date-time
                    example: "2025-11-15T15:30:00.000Z"
    EventWithUser:
      allOf:
        - $ref: '#/components/schemas/Event'
        - type: object
          properties:
            User:
              type: object
              properties:
                id:
                  type: integer
                  example: 1
                username:
                  type: string
                  example: "admin"
                email:
                  type: string
                  example: "admin@dawdle.com"
            Notifications:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: integer
                    example: 1
                  notif_time:
                    type: string
                    format: date-time
                    example: "2025-11-08T09:00:00.000Z"
            Completeds:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: integer
                    example: 1
                  completion_date:
                    type: string
                    format: date-time
                    example: "2025-11-15T15:30:00.000Z"
    Completed:
      type: object
      properties:
        id:
          type: integer
          example: 5
        event_id:
          type: integer
          example: 15
        completion_date:
          type: string
          format: date-time
          example: "2025-11-06T18:30:00.000Z"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []